+++ tests/integration/test_disk_read_write.py
@@ -1,11 +1,11 @@
-from openscm.highlevel import OpenSCMDataFrame
+from openscm.highlevel import ScmDataFrame
 
 
 def test_write_read_datafile(test_pd_df, tmp_path):
     tfile = str(tmp_path / "testfile.csv")
-    tdf = OpenSCMDataFrame(test_pd_df)
+    tdf = ScmDataFrame(test_pd_df)
 
     tdf.to_csv(tfile)
 
-    OpenSCMDataFrame(tfile)
+    ScmDataFrame(tfile)