+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 master
 ------
 
+- (`#83 <https://github.com/openclimatedata/openscm/pull/83>`_) Rename OpenSCMDataFrame to ScmDataFrame
 - (`#78 <https://github.com/openclimatedata/openscm/pull/78>`_) Added OpenSCMDataFrame
 - (`#35 <https://github.com/openclimatedata/openscm/pull/35>`_) Add units module
 - (`#44 <https://github.com/openclimatedata/openscm/pull/44>`_) Add timeframes module